Break Details component to parts
