+++ src/components/details/close-btn.js
@@ -1,0 +1,15 @@
+import AbstractComponent from '../abstract-component';
+
+export default class CloseBtn extends AbstractComponent {
+  _getTmpl() {
+    return (
+      `<div class="film-details__close">
+        <button class="film-details__close-btn" type="button">close</button>
+      </div>`
+    );
+  }
+
+  setClickHandler(handler) {
+    this._element.addEventListener(`click`, handler);
+  }
+}
